feat: add sanity.client.ts and sanity.preview.ts
This commit adds the sanity.client.ts and sanity.preview.ts files.

The sanity.client.ts file is used to create a client that can be used to fetch data from Sanity.

The sanity.preview.ts file is used to create a preview client that can be used to fetch data from Sanity.